% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_averages.R
\name{country_averages}
\alias{country_averages}
\title{Get country averages}
\usage{
country_averages(rankings, drop_countrys = TRUE)
}
\arguments{
\item{rankings}{Rankings created by \code{get_club_rankings}}

\item{drop_countrys}{Drop some miscellaneous teams that made it to Champions
or Europa country but whose domestic country is not included. Default is
true.}
}
\value{
A data frame of country rankings
}
\description{
Get average rankings per country
}
\details{
Takes the rankings you create with \code{get_club_rankings} and gets
  country averages.
}
\examples{
library(five38clubrankings)
country_averages(rankings_2018)
}
