% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five38clubrankings.R
\docType{package}
\name{five38clubrankings}
\alias{five38clubrankings}
\alias{five38clubrankings-package}
\title{five38clubrankings: Get the current rankings.}
\description{
This package just grabs the table located at
https://projects.fivethirtyeight.com/global-club-soccer-rankings/', and
additionally does some summary by league and country.  You can get the csv at
https://github.com/fivethirtyeight/data/tree/master/soccer-spi, but that
doesn't have country or one-week change. The main function is
\code{get_club_rankings}.
}
