% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_league_averages.R
\name{get_league_averages}
\alias{get_league_averages}
\title{Get league averages}
\usage{
get_league_averages(rankings, drop_leagues = TRUE)
}
\arguments{
\item{rankings}{Rankings created by \code{get_club_rankings}}

\item{drop_leagues}{Drop some miscellaneous teams that made it to Champions
or Europa League but whose domestic league is not included. Default is
TRUE.}
}
\value{
A data frame of league rankings
}
\description{
Get average rankings per league
}
\details{
Takes the rankings you create with \code{get_club_rankings} and gets
  league averages.
}
\examples{
library(five38clubrankings)
get_league_averages(rankings_2018)
}
