% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_2018.R
\docType{data}
\name{rankings_2018}
\alias{rankings_2018}
\title{Football club rankings}
\format{A data frame based on the csv or the website scrape.}
\source{
\url{https://projects.fivethirtyeight.com/global-club-soccer-rankings/}
}
\usage{
rankings_2018
}
\description{
Football/soccer club rankings as provided by 538.
}
\keyword{datasets}
