% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_2018_csv.R
\docType{data}
\name{rankings_2018_csv}
\alias{rankings_2018_csv}
\title{Football club rankings}
\format{A data frame based on the csv or the website scrape.}
\source{
\url{hhttps://github.com/fivethirtyeight/data/tree/master/soccer-spi/}
}
\usage{
rankings_2018_csv
}
\description{
Football club rankings
}
\details{
Football/soccer club rankings as provided by 538 around June of 2018.
}
\keyword{datasets}
