% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_club_rankings.R
\name{get_club_rankings}
\alias{get_club_rankings}
\title{Get the current rankings}
\usage{
get_club_rankings(csv = FALSE)
}
\arguments{
\item{csv}{See details.}
}
\value{
A data frame of the rankings
}
\description{
Download soccer club rankings as provided by 538.

Download soccer club rankings from 538
}
\details{
This uses the \href{https://github.com/hadley/rvest}{rvest} package to download football club rankings as
  provided by 538.  It has one argument, whether you want to use their API to
  get the `*.csv`, or not.  It is FALSE by default, as country and one-week
  change are available from the table on their website. The \code{\link[five38clubrankings]{country_averages}}
  function will not work if you use the csv.
}
\examples{
\dontrun{
get_club_rankings()
}
}
